Day 9, completed intcode computer
Involves using a map as memory instead of a mutable list, and using longs instead of ints for the program (longcode computer?)

+++ src/AdventOfCode2019/Day 2 Part 2.kt
@@ -6,11 +6,11 @@ fun main() {
 
     for (i in 0..99) {
         for (j in 0..99) {
-            instrs[1] = i
-            instrs[2] = j
+            instrs[1] = i.toLong()
+            instrs[2] = j.toLong()
 
             val res = runIntcodeProgram(instrs.toList())
-            if (res[0] == 19690720) {
+            if (res[0] == 19690720L) {
                 println(100 * i + j)
                 break
             }
